refactor: extract avatar management logic into dedicated ProfileImage model
